requirements.txt
flask
psycopg2-binary
jsonify
boto3